+++ srcs/geometry/tuples/init.c
@@ -6,13 +6,13 @@
 /*   By: lkrief <lkrief@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 02:34:33 by lkrief            #+#    #+#             */
-/*   Updated: 2023/02/08 09:38:11 by lkrief           ###   ########.fr       */
+/*   Updated: 2023/02/26 08:08:34 by lkrief           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "tuples.h"
 
-t_tuple	point(float x, float y, float z)
+t_tuple	point(double x, double y, double z)
 {
 	t_tuple	a;
 
@@ -23,7 +23,7 @@ t_tuple	point(float x, float y, float z)
 	return (a);
 }
 
-t_tuple	vector(float x, float y, float z)
+t_tuple	vector(double x, double y, double z)
 {
 	t_tuple	a;
 
@@ -34,7 +34,7 @@ t_tuple	vector(float x, float y, float z)
 	return (a);
 }
 
-t_tuple	tuple(float x, float y, float z, float w)
+t_tuple	tuple(double x, double y, double z, double w)
 {
 	t_tuple	a;
